Extract code into new files
